+++ example/basic/basic.dart
@@ -5,6 +5,6 @@ import 'package:ace/ace.dart' as ace;
 
 main() {
   var editor = ace.edit(querySelector('#editor'))
-  ..theme = new ace.Theme("ace/theme/monokai")
-  ..session.mode = new ace.Mode("ace/mode/dart");
+  ..theme = new ace.Theme('ace/theme/monokai')
+  ..session.mode = new ace.Mode('ace/mode/dart');
 }